+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+-   Replace RgbQuant with fast-average-color.
+-   Update deps.
+
 ## [0.3.0] - 2022-04-06
 
 ### Changed